cleaned code base
