+++ todolist/src/pages/Today/Task.module.css
@@ -78,7 +78,8 @@
   }
   
   .highlight {
-    background-color: yellow;
+    background-color: var(--ascent-color);
+
   }
   
   .invisibleInput {
